Remove remaining usage of `read`
